Fixed parser that incorrectly rejected floats containing a `+` (issue #349)

+++ include/ArduinoJson/Print.hpp
@@ -23,7 +23,7 @@ class Print {
   size_t print(const char* s) {
     size_t n = 0;
     while (*s) {
-      n += write(*s++);
+      n += write(static_cast<uint8_t>(*s++));
     }
     return n;
   }
